Move current example code to "simple" directory
