+++ examples/simple/config/slack/fixed_timer.yaml
