+++ examples/simple/config/slack/timer.yaml
