+++ examples/simple/plugins/count/props.go
