+++ examples/simple/plugins/echo/command.go
