+++ examples/simple/plugins/fixedtimer/props.go
