+++ examples/simple/plugins/guess/props.go
