+++ examples/simple/plugins/hello/props.go
