+++ examples/simple/plugins/morning/props.go
