+++ examples/simple/plugins/timer/props.go
