+++ examples/simple/plugins/todo/command.go
